updated book editing buttons, added user book viewer

+++ books/views.py
@@ -1,5 +1,6 @@
 from django.http import HttpResponse, HttpRequest
 from django.shortcuts import render, redirect, get_object_or_404
+from django.contrib.auth.decorators import login_required
 
 from .forms import BookForm
 from .models import Book
@@ -36,13 +37,37 @@ def list_books(request: HttpRequest):
         books = Book.objects.all().order_by("-title")
     return render(request, "books/home.html", context={"books": books})
 
+def list_user_books(request: HttpRequest, user_pk: int):
+    # trebuie sa listam cartile din baza de date
+    #accesare de carti:
+    #QuerySet
+
+    # request.GET este un dictionar care contine toate url params.
+    # "sort" este parametrul din urk care ne indica ce sortare facem.
+
+    sort = request.GET.get("sort")
+    books = Book.objects.filter(user_id=user_pk).order_by("pk")
+
+    # srt_b = sorted(list(books), key=lambda x: x.title.lower())
+
+    if sort == "asc":
+        books = Book.objects.filter(user_id=user_pk).order_by("title")
+    if sort == "desc":
+        books = Book.objects.filter(user_id=user_pk).order_by("-title")
+    return render(request, "books/home.html", context={"books": books})
+
+
+@login_required
 def create_book(request: HttpRequest):
     if request.method == "POST":
         #detaliile book-ului care au fost trimise de form folosind HTTP POST request, se afla in request.POST ca un dictionar.
-        book_instance = BookForm(request.POST)
-        if book_instance.is_valid():
+        form = BookForm(request.POST)
+        if form.is_valid():
             # aici se creeaza un book in baza de date!
-            book_instance.save()
+            book = form.save(commit=False)
+            book.user = request.user
+            book.save()
+
             return redirect("create_book")
     else:
         #in cazul asta, requestul poate fi GET, PUT, PATCH, DELETE, etc..
@@ -67,14 +92,16 @@ def update_book(request: HttpRequest, pk: int):
         return render(request, "books/update_book_form.html", context={"form": form})
 
 
-
+@login_required
 def delete_book(request: HttpRequest, pk: int):
     book = get_object_or_404(Book, pk=pk)
 
-    if request.method == "POST":
-        book.delete()
-        return redirect("home")
+    if request.user.pk == book.user.pk:
+        if request.method == "POST":
+            book.delete()
+            return redirect("home")
+        else:
+            return render(request, "books/book_confirm_delete.html", context={"book": book})
     else:
-        return render(request, "books/book_confirm_delete.html", context={"book": book})
+        return HttpResponse("You are not authorized to delete this book.")
 
-
